Create Plane class
Added Plane class with #land and #take_off methods which change the
plane's @landed state to true or false.  Calling #arrive and #depart
on instances of airport and passing an instance of plane as a parameter
will now call the plane's #land and #take_off methods to update state
accordingly.  Corresponding unit tests are passing.